seperate utility functions, change class name
